+++ CHANGES.md
@@ -7,3 +7,4 @@
 ## 2023-08-xx 0.0.0
 - Initial thing, using `arbitrary-dateparser` and `DateRangeParser` packages
 - Add German variant of `DateRangeParser`
+- Add `parse_single` method, for parsing a single point in time